requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
cycler==0.10.0
defusedxml==0.7.1
Django==3.1.7
django-crispy-forms==1.11.1
firebase-admin==4.5.2
flatbuffers==1.12
gast==0.3.3
gcloud==0.18.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.19.0
idna==2.10
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
msgpack==1.0.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.2
proto-plus==1.18.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
